+++ src/strategy/engine.js
@@ -6,6 +6,7 @@ import { vwapRevertStrategy } from './builtin/vwap-revert.js'
 import { spreadCaptureStrategy } from './builtin/spread-capture.js'
 import { bookImbalanceStrategy } from './builtin/book-imbalance.js'
 import { tapePressureStrategy } from './builtin/tape-pressure.js'
+import { rangeFadeStrategy } from './builtin/range-fade.js'
 import { safeInvoke } from './sandbox.js'
 
 /**
@@ -27,6 +28,7 @@ export const BUILTIN_STRATEGIES = Object.freeze([
   spreadCaptureStrategy,
   bookImbalanceStrategy,
   tapePressureStrategy,
+  rangeFadeStrategy,
   compositeStrategy,
 ])